+++ src/@types/index.d.ts
@@ -1,5 +1,8 @@
+export * from "./button";
 export * from "./canvas";
+export * from "./comment";
 export * from "./config";
+export * from "./cursor";
 export * from "./event";
 export * from "./fonts";
 export * from "./format.formatted";
